# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/content/injectables": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Injectables"
                ],
                "summary": "List injectables",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Locale for translations (default: es)",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ListInjectablesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/content/injectables/{injectableId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Injectables"
                ],
                "summary": "Get injectable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Injectable ID",
                        "name": "injectableId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.InjectableResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/content/templates/{templateId}/versions/{versionId}/injectables": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template Versions"
                ],
                "summary": "Add injectable to version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "templateId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "versionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Injectable data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.AddVersionInjectableRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.TemplateVersionInjectableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/content/templates/{templateId}/versions/{versionId}/injectables/{injectableId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template Versions"
                ],
                "summary": "Remove injectable from version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "templateId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "versionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Injectable ID",
                        "name": "injectableId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/system/injectables": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System - Injectables"
                ],
                "summary": "List system injectables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ListSystemInjectablesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/workspace/injectables": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Injectables"
                ],
                "summary": "List workspace injectables",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ListWorkspaceInjectablesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Injectables"
                ],
                "summary": "Create workspace injectable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Injectable data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.CreateWorkspaceInjectableRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.WorkspaceInjectableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/workspace/injectables/{injectableId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Injectables"
                ],
                "summary": "Get workspace injectable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Injectable ID",
                        "name": "injectableId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.WorkspaceInjectableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Injectables"
                ],
                "summary": "Update workspace injectable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Injectable ID",
                        "name": "injectableId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Injectable data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.UpdateWorkspaceInjectableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.WorkspaceInjectableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Injectables"
                ],
                "summary": "Delete workspace injectable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Injectable ID",
                        "name": "injectableId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.AddVersionInjectableRequest": {
            "type": "object",
            "required": [
                "injectableDefinitionId"
            ],
            "properties": {
                "defaultValue": {
                    "type": "string"
                },
                "injectableDefinitionId": {
                    "type": "string"
                },
                "isRequired": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.CreateWorkspaceInjectableRequest": {
            "type": "object",
            "required": [
                "defaultValue",
                "key",
                "label"
            ],
            "properties": {
                "defaultValue": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "key": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "label": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.FormatConfigResponse": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.GroupResponse": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                }
            }
        },
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.InjectableResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dataType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "formatConfig": {
                    "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.FormatConfigResponse"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isGlobal": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "sourceType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ListInjectablesResponse": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.GroupResponse"
                    }
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.InjectableResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ListSystemInjectablesResponse": {
            "type": "object",
            "properties": {
                "injectables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.SystemInjectableResponse"
                    }
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ListWorkspaceInjectablesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.WorkspaceInjectableResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.SystemInjectableResponse": {
            "type": "object",
            "properties": {
                "dataType": {
                    "type": "string"
                },
                "description": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "group": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isPublic": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.TemplateVersionInjectableResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "defaultValue": {
                    "type": "string"
                },
                "definition": {
                    "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.InjectableResponse"
                },
                "id": {
                    "type": "string"
                },
                "isRequired": {
                    "type": "boolean"
                },
                "templateVersionId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.UpdateWorkspaceInjectableRequest": {
            "type": "object",
            "properties": {
                "defaultValue": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "key": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "label": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.WorkspaceInjectableResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dataType": {
                    "type": "string"
                },
                "defaultValue": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "formatConfig": {
                    "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.FormatConfigResponse"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "sourceType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "string"
                }
            }
        },
...
    }
}